оптимизировал рендер: убрал перерендер компонента main при изменении фильтра, перерендер компонента moviesList при ховере на карточку фильма
